+++ wonton/support/test/test_structured_partitioner_par.cc
@@ -23,19 +23,19 @@ TEST(StructuredPartioner, Parallel) {
       // many partitions we are requesting).
 
       int seed = 42;
-      std::array<int, 3> ncells = {144, 256, 16};
+      std::array<int64_t, 3> ncells = {144, 256, 16};
       auto partlimits = Wonton::structured_partitioner<3>(16, ncells, 3, seed);
 
-      int limitsarray[96];  // flattened array
+      int64_t limitsarray[96];  // flattened array
       for (int i = 0; i < 16; i++)
         for (int j = 0; j < 3; j++) {
           limitsarray[i*6+j]   = partlimits[i][0][j];  // lower limit
           limitsarray[i*6+j+3] = partlimits[i][1][j];  // upper limit
         }
 
-      int limitsarray_all[384];
-      MPI_Allgather(limitsarray, 96, MPI_INT, limitsarray_all, 96, MPI_INT,
-                    MPI_COMM_WORLD);
+      int64_t limitsarray_all[384];
+      MPI_Allgather(limitsarray, 96, MPI_LONG_LONG, limitsarray_all, 96,
+                    MPI_LONG_LONG, MPI_COMM_WORLD);
 
       for (int i = 0; i < 16; i++)
         for (int j = 0; j < 6; j++)
